Refactor the RateLimitService to an ActionAttribute
