--- weather.station.server/weather.station.server/Services/IRateLimitService.cs
@@ -1,9 +1,0 @@
-namespace weather.station.server.Services
-{
-    public interface IRateLimitService
-    {
-        bool AllowRequest(string ip);
-
-        void RegisterRequest(string ip);
-    }
-}
