+++ weather.station.server/weather.station.server/Startup.cs
@@ -13,7 +13,6 @@ using Microsoft.Extensions.Options;
 using Microsoft.EntityFrameworkCore;
 using weather.station.server.Data;
 using Pomelo.EntityFrameworkCore.MySql.Infrastructure;
-using weather.station.server.Services;
 using Microsoft.AspNetCore.Http;
 
 namespace weather.station.server
@@ -49,12 +48,6 @@ namespace weather.station.server
                 services.AddDbContext<WeatherStationServerContext>(options =>
                     options.UseMySql(connectionString));
             }
-
-            // Register services
-            services.AddSingleton<IRateLimitService, RateLimitService>();
-
-            // Configure custom services
-            services.Configure<RateLimitServiceOptions>(Configuration.GetSection("RateLimit"));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
